+++ apps/uiservice/src/components/header/index.tsx
@@ -63,8 +63,8 @@ export default function Header() {
                 </Button>
               </DropdownMenuTrigger>
               <DropdownMenuContent align="end">
-                <DropdownMenuItem onClick={(): void => router.push("/account/user")}>Account Settings</DropdownMenuItem>
-                <DropdownMenuItem onClick={(): void => router.push("/account/organization")}>Organization</DropdownMenuItem>
+                <DropdownMenuItem onClick={(): void => router.push("/settings/user")}>Settings</DropdownMenuItem>
+                <DropdownMenuItem onClick={(): void => router.push("/settings/organization")}>Organization</DropdownMenuItem>
                 <DropdownMenuItem onClick={(): void => router.push("/")}>Home Page</DropdownMenuItem>
                 <DropdownMenuItem onClick={signOut}>Sign Out</DropdownMenuItem>
               </DropdownMenuContent>